+++ tp_router/example/lib/pages/dashboard/dashboard_shell.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+import 'package:tp_router/tp_router.dart';
+import 'analytics_page.dart';
+import 'reports_page.dart';
+import 'dashboard_overview_page.dart';
+import '../../tp_router.g.dart';
+
+@TpShellRoute(
+  children: [DashboardOverviewPage, AnalyticsPage, ReportsPage],
+)
+class DashboardShell extends StatelessWidget {
+  final Widget child;
+
+  const DashboardShell({
+    required this.child,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Dashboard'),
+      ),
+      body: child,
+    );
+  }
+}